Move examples to separate module
